+++ .babelrc.js
@@ -1,23 +1,25 @@
 const { NODE_ENV } = process.env
 
 module.exports = {
-    presets: [
-        [
-            '@babel/env',
-            {
-                targets: {
-                    browsers: ['ie >= 11']
-                },
-                exclude: ['transform-async-to-generator', 'transform-regenerator'],
-                modules: false,
-                loose: true
-            }
-        ]
+  presets: [
+    '@babel/typescript',
+    [
+      '@babel/env',
+      {
+        targets: {
+          browsers: ['ie >= 11'],
+        },
+        exclude: ['transform-async-to-generator', 'transform-regenerator'],
+        modules: false,
+        loose: true,
+      },
     ],
-    plugins: [
-        '@babel/plugin-proposal-class-properties',
-        // don't use `loose` mode here - need to copy symbols when spreading
-        '@babel/proposal-object-rest-spread',
-        NODE_ENV === 'test' && '@babel/transform-modules-commonjs'
-    ].filter(Boolean)
+  ],
+  plugins: [
+    '@babel/proposal-private-methods',
+    '@babel/proposal-class-properties',
+    // don't use `loose` mode here - need to copy symbols when spreading
+    '@babel/proposal-object-rest-spread',
+    NODE_ENV === 'test' && '@babel/transform-modules-commonjs',
+  ].filter(Boolean),
 }