+++ jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    testPathIgnorePatterns: ['/node_modules/'],
-    moduleNameMapper: {
-        '^src(.*)$': '<rootDir>/src$1',
-    },
-}